(documentation)(refactor): Added documentaion to RootStackParamList and moved it into index.tsx
