+++ src/screens/SignInScreen.tsx
@@ -11,7 +11,7 @@ import db, { auth, userRef } from "../../config/firebase";
 import { doc, setDoc } from "firebase/firestore";
 import { useNavigation } from "@react-navigation/native";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../../App";
+import { RootStackParamList } from "../navigation/index";
 import {
   Container,
   LoginText,
